Added rotated hamiltonian method
